Move RS0013 to the new location. Mechanical move.
