+++ src/Roslyn.Diagnostics.Analyzers/Core/DoNotInvokeDiagnosticDescriptor.cs
@@ -1,38 +1,68 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Collections.Immutable;
-using Microsoft.CodeAnalysis.Diagnostics;
+using Analyzer.Utilities;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Roslyn.Diagnostics.Analyzers
-{                   
-    /// <summary>
-    /// RS0013: Do not invoke Diagnostic.Descriptor
-    /// </summary>
-    public abstract class DoNotInvokeDiagnosticDescriptorAnalyzer : DiagnosticAnalyzer
+{
+    public abstract class DiagnosticDescriptorAccessAnalyzer<TSyntaxKind, TMemberAccessExpressionSyntax> : DiagnosticAnalyzer
+        where TSyntaxKind : struct
+        where TMemberAccessExpressionSyntax : SyntaxNode
     {
-        internal const string RuleId = "RS0013";
+        private static readonly string s_diagnosticTypeFullName = typeof(Diagnostic).FullName;
 
         private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotInvokeDiagnosticDescriptorTitle), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
-        
         private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotInvokeDiagnosticDescriptorMessage), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
         private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotInvokeDiagnosticDescriptorDescription), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
-        
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                             s_localizableTitle,
-                                                                             s_localizableMessage,
-                                                                             DiagnosticCategory.Performance,
-                                                                             DiagnosticSeverity.Warning,
-                                                                             isEnabledByDefault: false,
-                                                                             description: s_localizableDescription,
-                                                                             helpLinkUri: null,     // TODO: add MSDN url
-                                                                             customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
+        internal static readonly DiagnosticDescriptor DoNotRealizeDiagnosticDescriptorRule = new DiagnosticDescriptor(
+            RoslynDiagnosticIds.DoNotAccessDiagnosticDescriptorRuleId,
+            s_localizableTitle,
+            s_localizableMessage,
+            "Performance",
+            DiagnosticSeverity.Warning,
+            isEnabledByDefault: false,
+            description: s_localizableDescription,
+            customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override void Initialize(AnalysisContext analysisContext)
-        { 
-            
+        public sealed override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(DoNotRealizeDiagnosticDescriptorRule);
+
+        public sealed override void Initialize(AnalysisContext context)
+        {
+            context.RegisterSyntaxNodeAction(AnalyzeNode, SimpleMemberAccessExpressionKind);
+        }
+
+        protected abstract TSyntaxKind SimpleMemberAccessExpressionKind { get; }
+
+        protected abstract SyntaxNode GetLeftOfMemberAccess(TMemberAccessExpressionSyntax memberAccess);
+        protected abstract SyntaxNode GetRightOfMemberAccess(TMemberAccessExpressionSyntax memberAccess);
+        protected abstract bool IsThisOrBaseOrMeOrMyBaseExpression(SyntaxNode node);
+
+        protected void AnalyzeNode(SyntaxNodeAnalysisContext context)
+        {
+            var memberAccess = (TMemberAccessExpressionSyntax)context.Node;
+            var right = GetRightOfMemberAccess(memberAccess);
+            if (right.ToString() != nameof(Diagnostic.Descriptor))
+            {
+                return;
+            }
+
+            var left = GetLeftOfMemberAccess(memberAccess);
+            var leftType = context.SemanticModel.GetTypeInfo(left).Type;
+            if (leftType != null && leftType.ToDisplayString() == s_diagnosticTypeFullName && !IsThisOrBaseOrMeOrMyBaseExpression(left))
+            {
+                var nameOfMember = string.Empty;
+                var parentMemberAccess = memberAccess.Parent as TMemberAccessExpressionSyntax;
+                if (parentMemberAccess != null)
+                {
+                    var member = GetRightOfMemberAccess(parentMemberAccess);
+                    nameOfMember = " '" + member.ToString() + "'";
+                }
+
+                memberAccess.CreateDiagnostic(DoNotRealizeDiagnosticDescriptorRule, nameof(Diagnostic.Descriptor), nameof(Diagnostic), nameOfMember);
+            }
         }
     }
 }
