+++ src/Roslyn.Diagnostics.Analyzers/Core/RoslynDiagnosticsAnalyzersResources.resx
@@ -170,7 +170,7 @@
     <value>Accessing the Descriptor property of Diagnostic in compiler layer leads to unnecessary string allocations for fields of the descriptor that are not utilized in command line compilation. Hence, you should avoid accessing the Descriptor of the compiler diagnostics here. Instead you should directly access these properties off the Diagnostic type.</value>
   </data>
   <data name="DoNotInvokeDiagnosticDescriptorMessage" xml:space="preserve">
-    <value>Do not invoke Diagnostic.Descriptor</value>
+    <value>Do not invoke property '{0}' on type '{1}', instead directly access the required member{2} on '{1}'</value>
   </data>
   <data name="SymbolDeclaredEventMustBeGeneratedForSourceSymbolsTitle" xml:space="preserve">
     <value>SymbolDeclaredEvent must be generated for source symbols</value>
